requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
sqlite3
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
requests>=2.31.0
sounddevice>=0.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
thread6>=0.4.0